requirements.txt
asgiref==3.11.1
Django==5.2
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
